# server/openapi.json
{
	"components": {
		"schemas": {
...
			"AuthStatusDTO": {
				"description": "AuthStatusDTO schema",
				"properties": {
					"setup_required": {
						"type": "boolean"
					}
				},
				"required": [
					"setup_required"
				],
				"type": "object"
			},
			"ChangePasswordInput": {
				"description": "ChangePasswordInput schema",
				"properties": {
					"current_password": {
						"type": "string"
					},
					"new_password": {
						"type": "string"
					}
				},
				"required": [
					"current_password",
					"new_password"
				],
				"type": "object"
			},
...
			"CredentialsInput": {
				"description": "CredentialsInput schema",
				"properties": {
					"password": {
						"type": "string"
					}
				},
				"required": [
					"password"
				],
				"type": "object"
			},
...
			"ErrorItem": {
				"properties": {
					"more": {
						"additionalProperties": {
							"description": "Additional information about the error"
						},
						"description": "Additional information about the error",
						"type": [
							"object",
							"null"
						]
					},
					"name": {
						"description": "For example, name of the parameter that caused the error",
						"type": "string"
					},
					"reason": {
						"description": "Human readable error message",
						"type": "string"
					}
				},
				"required": [
					"name",
					"reason"
				],
				"type": "object"
			},
			"HTTPError": {
				"description": "HTTPError schema",
				"properties": {
					"detail": {
						"description": "Human readable error message",
						"type": "string"
					},
					"errors": {
						"items": {
							"$ref": "#/components/schemas/ErrorItem"
						},
						"type": [
							"array",
							"null"
						]
					},
					"instance": {
						"type": "string"
					},
					"status": {
						"description": "HTTP status code",
						"example": 403,
						"type": "integer"
					},
					"title": {
						"description": "Short title of the error",
						"type": "string"
					},
					"type": {
						"description": "URL of the error type. Can be used to lookup the error in a documentation",
						"type": "string"
					}
				},
				"type": "object"
			},
...
			"OK": {
				"description": "OK schema",
				"properties": {
					"ok": {
						"type": "boolean"
					}
				},
				"required": [
					"ok"
				],
				"type": "object"
			},
...
			"Status": {
				"description": "Status schema",
				"properties": {
					"current_version": {
						"type": "string"
					},
					"error": {
						"type": "string"
					},
					"last_checked": {
						"format": "date-time",
						"type": "string"
					},
					"latest_version": {
						"type": "string"
					},
					"state": {
						"type": "string"
					},
					"update_available": {
						"type": "boolean"
					}
				},
				"required": [
					"current_version",
					"last_checked",
					"latest_version",
					"state",
					"update_available"
				],
				"type": "object"
			},
			"TokenDTO": {
				"description": "TokenDTO schema",
				"properties": {
					"token": {
						"type": "string"
					}
				},
				"required": [
					"token"
				],
				"type": "object"
			},
...
		}
	},
	"info": {
...
		"title": "OpenAPI",
		"version": "0.0.1"
	},
	"openapi": "3.1.0",
	"paths": {
		"/api/auth/login": {
			"post": {
				"description": "#### Controller: \n\n`github.com/bwees/mundus/server/internal/webapi.registerAuth.func3`\n\n#### Middlewares:\n\n- `github.com/go-fuego/fuego.defaultLogger.middleware`\n\n---\n\n",
				"operationId": "login",
				"parameters": [
					{
						"in": "header",
						"name": "Accept",
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"content": {
						"*/*": {
							"schema": {
								"$ref": "#/components/schemas/CredentialsInput"
							}
						}
					},
					"description": "Request body for webapi.CredentialsInput",
					"required": true
				},
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TokenDTO"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/TokenDTO"
								}
							}
						},
						"description": "OK"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							}
						},
						"description": "Bad Request _(validation or deserialization error)_"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							}
						},
						"description": "Internal Server Error _(panics)_"
					}
				},
				"summary": "func3",
				"tags": [
					"api"
				]
			}
		},
		"/api/auth/password": {
			"post": {
				"description": "#### Controller: \n\n`github.com/bwees/mundus/server/internal/webapi.registerAuth.func4`\n\n#### Middlewares:\n\n- `github.com/go-fuego/fuego.defaultLogger.middleware`\n\n---\n\n",
				"operationId": "changePassword",
				"parameters": [
					{
						"in": "header",
						"name": "Accept",
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"content": {
						"*/*": {
							"schema": {
								"$ref": "#/components/schemas/ChangePasswordInput"
							}
						}
					},
					"description": "Request body for webapi.ChangePasswordInput",
					"required": true
				},
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TokenDTO"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/TokenDTO"
								}
							}
						},
						"description": "OK"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							}
						},
						"description": "Bad Request _(validation or deserialization error)_"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							}
						},
						"description": "Internal Server Error _(panics)_"
					}
				},
				"summary": "func4",
				"tags": [
					"api"
				]
			}
		},
		"/api/auth/setup": {
			"post": {
				"description": "#### Controller: \n\n`github.com/bwees/mundus/server/internal/webapi.registerAuth.func2`\n\n#### Middlewares:\n\n- `github.com/go-fuego/fuego.defaultLogger.middleware`\n\n---\n\n",
				"operationId": "setupAuth",
				"parameters": [
					{
						"in": "header",
						"name": "Accept",
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"content": {
						"*/*": {
							"schema": {
								"$ref": "#/components/schemas/CredentialsInput"
							}
						}
					},
					"description": "Request body for webapi.CredentialsInput",
					"required": true
				},
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TokenDTO"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/TokenDTO"
								}
							}
						},
						"description": "OK"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							}
						},
						"description": "Bad Request _(validation or deserialization error)_"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							}
						},
						"description": "Internal Server Error _(panics)_"
					}
				},
				"summary": "func2",
				"tags": [
					"api"
				]
			}
		},
		"/api/auth/status": {
			"get": {
				"description": "#### Controller: \n\n`github.com/bwees/mundus/server/internal/webapi.registerAuth.func1`\n\n#### Middlewares:\n\n- `github.com/go-fuego/fuego.defaultLogger.middleware`\n\n---\n\n",
				"operationId": "getAuthStatus",
				"parameters": [
					{
						"in": "header",
						"name": "Accept",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthStatusDTO"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/AuthStatusDTO"
								}
							}
						},
						"description": "OK"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							}
						},
						"description": "Bad Request _(validation or deserialization error)_"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							}
						},
						"description": "Internal Server Error _(panics)_"
					}
				},
				"summary": "func1",
				"tags": [
					"api"
				]
			}
		},
...
		"/api/update": {
			"get": {
				"description": "#### Controller: \n\n`github.com/bwees/mundus/server/internal/webapi.registerUpdate.func1`\n\n#### Middlewares:\n\n- `github.com/go-fuego/fuego.defaultLogger.middleware`\n\n---\n\n",
				"operationId": "getUpdate",
				"parameters": [
					{
						"in": "header",
						"name": "Accept",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Status"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/Status"
								}
							}
						},
						"description": "OK"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							}
						},
						"description": "Bad Request _(validation or deserialization error)_"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							}
						},
						"description": "Internal Server Error _(panics)_"
					}
				},
				"summary": "func1",
				"tags": [
					"api"
				]
			}
		},
		"/api/update/apply": {
			"post": {
				"description": "#### Controller: \n\n`github.com/bwees/mundus/server/internal/webapi.registerUpdate.func3`\n\n#### Middlewares:\n\n- `github.com/go-fuego/fuego.defaultLogger.middleware`\n\n---\n\n",
				"operationId": "applyUpdate",
				"parameters": [
					{
						"in": "header",
						"name": "Accept",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/OK"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/OK"
								}
							}
						},
						"description": "OK"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							}
						},
						"description": "Bad Request _(validation or deserialization error)_"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							}
						},
						"description": "Internal Server Error _(panics)_"
					}
				},
				"summary": "func3",
				"tags": [
					"api"
				]
			}
		},
		"/api/update/check": {
			"post": {
				"description": "#### Controller: \n\n`github.com/bwees/mundus/server/internal/webapi.registerUpdate.func2`\n\n#### Middlewares:\n\n- `github.com/go-fuego/fuego.defaultLogger.middleware`\n\n---\n\n",
				"operationId": "checkUpdate",
				"parameters": [
					{
						"in": "header",
						"name": "Accept",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Status"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/Status"
								}
							}
						},
						"description": "OK"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							}
						},
						"description": "Bad Request _(validation or deserialization error)_"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							}
						},
						"description": "Internal Server Error _(panics)_"
					}
				},
				"summary": "func2",
				"tags": [
					"api"
				]
			}
		}
	},
...
}